rapozcode-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-genai==1.23.0
python-dotenv==1.0.0
requests==2.31.0 